Move 'patch_openai_client' tests to 'clients/python/tests'
This avoids the need for 'openai-python' to depend on the
'tensorzero' python package. I've also adjusted one of the tests
for deprecation behvaior to use a patched client, to avoid triggering
deprecation warnings in the gatewat
