+++ clients/openai-python/pyproject.toml
@@ -4,10 +4,9 @@ version = "0.0.0"
 readme = "README.md"
 requires-python = ">=3.9"
 dependencies = [
-    "tensorzero @ file://${PROJECT_ROOT}/../python",
     "openai",
     "pydantic",
 ]
 
 [dependency-groups]
-dev = ["pytest", "pytest-asyncio"]
+dev = ["pytest", "pytest-asyncio", "uuid_utils"]
